Move BlockchainClient::Base to Helpers module
